chore: 폴더 구조 및 파일 네임 변경
